requirements.txt
requests
pyspark
apache-airflow